+++ tests/cases/compiler/declarationEmitInferedTypeAlias2.ts
@@ -1,4 +1,4 @@
-// @declaration: true
+﻿// @declaration: true
 // @skipDefaultLibCheck: true
 
 // @Filename: 0.ts
@@ -9,5 +9,8 @@
 export { }
 
 // @Filename: 1.ts
-var x = "hi" || 5;
-export default x;
+let v = "str" || true;
+function bar () {
+    return v;
+}
+export { v, bar }
