+++ tests/cases/compiler/declarationEmitInferedTypeAlias4.ts
@@ -1,0 +1,7 @@
+﻿// @declaration: true
+
+function f<A>() {
+    type Foo<T> = T | { x: Foo<T> };
+    var x: Foo<A[]>;
+    return x;
+}
