+++ tests/cases/compiler/declarationEmitInferedTypeAlias5.ts
@@ -2,12 +2,11 @@
 // @skipDefaultLibCheck: true
 
 // @Filename: 0.ts
-{
-    type Data = string | boolean;
-    let obj: Data = true;
-}
-export { }
+export type Data = string | boolean;
+let obj: Data = true;
 
 // @Filename: 1.ts
+import * as Z from "./0"
+//let v2: Z.Data;
 let v = "str" || true;
 export { v }
