+++ tests/cases/compiler/declarationEmitInferedTypeAlias6.ts
