+++ tests/cases/compiler/declarationEmitInferedTypeAlias8.ts
@@ -1,0 +1,8 @@
+// @declaration: true
+
+type Foo<T> = T | { x: Foo<T> };
+var x: Foo<number[]>;
+
+function returnSomeGlobalValue() {
+    return x;
+}
