+++ tests/cases/compiler/declarationEmitInferedTypeAlias9.ts
@@ -1,0 +1,8 @@
+// @declaration: true
+
+type Foo<T> = T | { x: Foo<T> };
+var x: Foo<number[]>;
+
+export function returnSomeGlobalValue() {
+    return x;
+}
